# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  customer: "customer",
  translator: "translator",
  admin: "admin",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserPublicRole =
  (typeof UserPublicRole)[keyof typeof UserPublicRole];

export const UserPublicRole = {
  customer: "customer",
  translator: "translator",
  admin: "admin",
} as const;

export interface UserPublic {
  id: number;
  email: string;
  role: UserPublicRole;
}

export interface AuthResponse {
  token: string;
  user: UserPublic;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  customer: "customer",
  translator: "translator",
  admin: "admin",
} as const;

export interface CreateUserRequest {
  email: string;
  role?: CreateUserRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  translator: "translator",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface CreateProjectRequest {
  title: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  created: "created",
  quoted: "quoted",
  approved: "approved",
  matched: "matched",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  userId: number;
  title: string;
  status: ProjectStatus;
  createdAt: string;
}

export interface CreateQuoteRequest {
  projectId: number;
  /** @minimum 0 */
  price: number;
}

export type QuoteStatus = (typeof QuoteStatus)[keyof typeof QuoteStatus];

export const QuoteStatus = {
  pending: "pending",
  sent: "sent",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Quote {
  id: number;
  projectId: number;
  price: string;
  status: QuoteStatus;
  createdAt: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  waiting: "waiting",
  assigned: "assigned",
  working: "working",
  done: "done",
} as const;

export interface Task {
  id: number;
  projectId: number;
  translatorId: number;
  status: TaskStatus;
  createdAt: string;
}

export type TaskWithProjectStatus =
  (typeof TaskWithProjectStatus)[keyof typeof TaskWithProjectStatus];

export const TaskWithProjectStatus = {
  waiting: "waiting",
  assigned: "assigned",
  working: "working",
  done: "done",
} as const;

export interface TaskWithProject {
  id: number;
  projectId: number;
  translatorId: number;
  status: TaskWithProjectStatus;
  createdAt: string;
  projectTitle?: string | null;
  projectStatus?: string | null;
}

export type LogEntityType = (typeof LogEntityType)[keyof typeof LogEntityType];

export const LogEntityType = {
  project: "project",
  quote: "quote",
  task: "task",
} as const;

export interface Log {
  id: number;
  entityType: LogEntityType;
  entityId: number;
  action: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProjectsParams = {
  userId?: number;
};

export type ListTasksParams = {
  translatorId?: number;
};

export type ListLogsParams = {
  entityType?: ListLogsEntityType;
  entityId?: number;
};

export type ListLogsEntityType =
  (typeof ListLogsEntityType)[keyof typeof ListLogsEntityType];

export const ListLogsEntityType = {
  project: "project",
  quote: "quote",
  task: "task",
} as const;