# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.5.0
  description: 통번역 플랫폼 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: projects
    description: Project operations
  - name: quotes
    description: Quote operations
  - name: tasks
    description: Task operations
  - name: logs
    description: Log operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered and token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (no password, legacy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects:
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (customer only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (wrong role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects (optional userId filter)
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{id}/match:
    post:
      operationId: matchProject
      tags: [projects]
      summary: Match a translator to an approved project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Task created and translator matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Project not approved or already matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project or translator not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes:
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a quote for a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}/approve:
    post:
      operationId: approveQuote
      tags: [quotes]
      summary: Approve a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Already approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optional translatorId filter)
      parameters:
        - name: translatorId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks with project info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskWithProject"
  /tasks/{id}/start:
    patch:
      operationId: startTask
      tags: [tasks]
      summary: Start a task (translator only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not assigned to this task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task (translator only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not assigned to this task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List all event logs
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
            enum: [project, quote, task]
        - name: entityId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Log"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, translator, admin]
          default: customer
      required:
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required:
        - token
        - user
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [customer, translator, admin]
      required:
        - id
        - email
        - role
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [customer, translator, admin]
          default: customer
      required:
        - email
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [customer, translator, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [created, quoted, approved, matched, in_progress, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - status
        - createdAt
    CreateQuoteRequest:
      type: object
      properties:
        projectId:
          type: integer
        price:
          type: number
          minimum: 0
      required:
        - projectId
        - price
    Quote:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        price:
          type: string
        status:
          type: string
          enum: [pending, sent, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - price
        - status
        - createdAt
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        translatorId:
          type: integer
        status:
          type: string
          enum: [waiting, assigned, working, done]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - translatorId
        - status
        - createdAt
    TaskWithProject:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        translatorId:
          type: integer
        status:
          type: string
          enum: [waiting, assigned, working, done]
        createdAt:
          type: string
          format: date-time
        projectTitle:
          type: string
          nullable: true
        projectStatus:
          type: string
          nullable: true
      required:
        - id
        - projectId
        - translatorId
        - status
        - createdAt
    Log:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
          enum: [project, quote, task]
        entityId:
          type: integer
        action:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entityType
        - entityId
        - action
        - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error