# lib/api-zod/src/generated/types/createQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */

export interface CreateQuoteRequest {
  projectId: number;
  /** @minimum 0 */
  price: number;
}