# lib/api-zod/src/generated/types/log.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { LogEntityType } from "./logEntityType";

export interface Log {
  id: number;
  entityType: LogEntityType;
  entityId: number;
  action: string;
  createdAt: Date;
}