# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */

export interface LoginRequest {
  email: string;
  password: string;
}