# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { QuoteStatus } from "./quoteStatus";

export interface Quote {
  id: number;
  projectId: number;
  price: string;
  status: QuoteStatus;
  createdAt: Date;
}