# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterRequestRole;
}