# lib/api-zod/src/generated/types/taskWithProject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { TaskWithProjectStatus } from "./taskWithProjectStatus";

export interface TaskWithProject {
  id: number;
  projectId: number;
  translatorId: number;
  status: TaskWithProjectStatus;
  createdAt: Date;
  projectTitle?: string | null;
  projectStatus?: string | null;
}