# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  role: UserRole;
  createdAt: Date;
}