# lib/api-zod/src/generated/types/userPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 통번역 플랫폼 API
 * OpenAPI spec version: 0.5.0
 */
import type { UserPublicRole } from "./userPublicRole";

export interface UserPublic {
  id: number;
  email: string;
  role: UserPublicRole;
}